Select full metadata when generating doc proxies
Motivation
----------
When saving documents back during change tracking, we want to be able to
use the CAS as well as the document ID.

Modifications
-------------
Change proxy generation to select the entire metadata object, including
CAS, rather than just the document ID.  This is then tracked in the document
node.

Results
-------
CAS is now available for change tracking to prevent collisions.