Prevent serialization of metadata for proxied documents
Motivation
----------
Proxied documents have extra properties containing document metadata.
These properties should not be serialized back to Couchbase if they are
written.

Modifications
-------------
Changed deserialization approach for the __metadata attribute.  Instead of
expecting the deserializer to read directly into ITrackedDocumentNode, we
add the property directly on the proxy.  This property only has a setter,
no getter, so it won't be serialized.  The setter has a generated IL
implementation that writes the value to the ITrackedDocumentNode property.

Added IgnoreDataMember attributes to all ITrackedDocumentNode properties,
including metadata.

Changed "__metadata" property on ITrackedDocumentNode to "Metadata", as
special naming is no longer required.

Also discovered that virtual properties were not being output by
Newtonsoft for proxy objects.  Fixed this by ensuring the the proxy also
overrides the getter methods in DocumentProxyGenerationHook.

Results
-------
Metadata is no longer serialized with the document if it was proxied when
read.