Support for Couchbase 4.5 index joins
Motivation
----------
Couchbase 4.5 allows for flexible commutative joins, aka index joins.  For
more information, see
http://developer.couchbase.com/documentation/server/4.5-dp/flexible-join-n1ql.html.

Modifications
-------------
Added Versioning namespace, which defines a provider for determining and
caching the version number for a cluster.  This includes a suite of unit
and integration tests.

Updated the query generation system to output index joins when detected in
the LINQ query, but only if the cluster is on version 4.5 or later.  If
the cluster isn't on version 4.5, a NotSupportedException is thrown which
is consistent with previous behaviors.

Results
-------
Index joins can now be used against 4.5 clusters.

The behavior is largely unaffected for clients running against 4.0 and 4.1
clusters.  Differences include additional logs, a slight change to an
exception message, and a miniscule delay on the first query while the
version number of the cluster is determined.