Issue #153: Support for controlling array subquery extent names
Motivation
----------
To use the new array index feature in Couchbase Server 4.5, the extent
names in array subqueries must exactly match the extent names used when
creating the array index.  This is incompatible with the current approach
of automatically generated extent names.

Modifications
-------------
Add IEnumerable<T> extension AsQueryable(extentName).  This overloads the
existing .Net AsQueryable() method.

When found, this expression node causes the ExtentNameProvider to use the
hard-coded extent name instead of an automatically generated name.

Created unit and integration tests for the new functionality.

Results
-------
It is now possible for SDK consumers to programatically control the extent
name used for array subqueries.  The consumer is responsible for ensuring
that they don't create naming conflicts in the resulting N1QL query.