Perform consistency checks when calling SubmitChanges
Motivation
----------
It is generally expected behavior from an SDK like this to prevent
modifications to documents that have already been modified by other
processes, unless expressly overridden.

Modifications
-------------
Added SaveOptions class and optional parameter to SubmitChanges to allow
consistency checking to be disabled.  Defaults to enabled.

Updated SubmitChanges logic to pass the CAS from DocumentMetadata if
consistency checks are enabled.

Created CouchbaseConsistencyException, inherited from
CouchbaseWriteException, to differentiate between consistency failures and
other types of write failures.

Created unit and integration tests to cover the new functionality.

Results
-------
Consistency checking is now the default behavior for change tracking.  It
behaves in a simple and easy to understand manner for the consumer.