Make Save behavior on BucketContext more logical during change tracking
Motivation
----------
When using Save on BucketContext with documents other than new documents
it doesn't behave in a logical manner.

Modifications
-------------
Move RegisterChangeTracking call into Track method on BucketContext for
improved consistency.  Also ensure that Untrack calls
UnregisterChangeTracking.

When a document is removed then readded to the context, treat as a
modification instead of as a delete.

Calling Save on a document already flagged as modified has no effect.

Calling Save on an already tracked document flags the document as
modified and dirty so that it posts to Couchbase on SubmitChanges.

Results
-------
Behavior of Save is more consistent with what a consumer would expect.