Issue #191: Allow ExecuteAsync to be used with in-memory collections
Motivation
----------
If working with a mock IBucketContext, calling Query<T> will normally
return an in-memory collection converted to an IQueryable.  In this
case, it will be useful if ExecuteAsync simply processes the in-memory
collection synchronously for the test rather than throwing an exception.

Modifications
-------------
Detect an in-memory collection (based on EnumerableQuery) and execute
synchronously.

Results
-------
Consumers mocking IBucketContext may now use ExecuteAsync against
in-memory collections.