Issue #199: Add option to suppress document filters
Motivation
----------
When performing LEFT JOINs, filters which are automatically injected via
the DocumentFilterManager can produce unexpected behaviors.  This is due
to a N1QL limitation that you cannot apply filters before a JOIN, only
after the JOIN is complete.

Modifications
-------------
Added BucketContext.Query overload that allows the user to suppress the
filters, such as the DocumentTypeFilter.

Added unit tests for all variants to ensure proper behavior.

Results
-------
When performing a LEFT JOIN, the user can apply
BucketQueryOptions.SuppressFilters to the right hand side of the join.
This will prevent the LINQ expressions which cannot be translated to
N1QL from being applied to the query.