Use new SELECT RAW syntax to extract scalar results on Server 5.0
Motivation
----------
Couchbase Server 5.0 offers a new SELECT RAW syntax which can extract
raw scalars in the result set, rather than the JSON objects always
returned by SELECT.  The current workaround is to use this syntax:

`SELECT Extent1.property as result FROM bucket as Extent1`

The `result` attribute is then extracted from the object for each row
returned.  The new syntax can be:

`SELECT RAW Extent1.property FROM bucket as Extent1`

This syntax doesn't require the wrapping object with a `result`
attribute.  This reduces data on the wire, heap allocations during
deserialization, and the number of operations to be performed processing
each result row.

This support will also be useful in future commits to implement array
subqueries, another new 5.0 feature.

Modifications
-------------
Use SELECT RAW wherever applicable instead of result extraction, so long
as the cluster is on version 5.0 or greater.

Test each of these circumstances to ensure property queries are
generated on pre-5.0 and post-5.0 servers.

Results
-------
The following types of queries now use SELECT RAW for performance if the
cluster version is 5.0 or greater:

- Selecting a document POCO directly without projection, so long as
change tracking is not enabled
- Selecting scalar properties directly
- Subqueries returning scalar properties directly
- Standalone aggregates