Issue #225: Support for Dictionary<TKey, TValue>.ContainsKey(TKey)
Motivation
----------
Allow queries to test for the presense of a particular key on an object
which is serialized from a Dictionary implementation.

Modifications
-------------
Implement DictionaryContainsKeyMethodCallTranslator, along with matching
unit tests.

Fix a bug in the DefaultMethodCallTranslatorProvider related to
detecting interface method translations from a class.

Results
-------
ContainsKey method calls on dictionaries may now be used in queries.