Issue #225: List dictionary keys and values
Motivation
----------
Allow the enumeration of dictionary keys and values within N1QL queries.

Modifications
-------------
May Dictionary.Keys to the OBJECT_NAMES function, and Dictionary.Values
to the OBJECT_VALUES function.

Create supporting unit tests, and also fill in some missing integration
tests for all supported dictionary functions.

Results
-------
Dictionary keys and values can now be enumerated within N1QL queries.

Note: To return one of these lists in the select projection, it is
necessary to call .ToList() on them.  The type returned by is not
deserializable because it has a private constructor.  This is only
necessary when returning the list in the select projection, not for
using the list in other clauses.