Issue #229: Support for async query cancellation
Motivation
----------
Allow long running LINQ queries executing asynchronously to be
cancelled if the result is no longer required.  For example, if an
incoming HTTP request has its connection dropped before the query
completes.

Modifications
-------------
Added overloads to ExecuteAsync to accept a CancellationToken, defaulted
to none.

Updated backing methods to accept a CancellationToken and pass it on the
the Couchbase SDK.

Results
-------
Async queries may now be cancelled.