Issue #223: DateTime fields as Unix milliseconds
Motivation
----------
Support date/time comparisons and functions on values persisted as
milliseconds since the Unix epoch.

Modifications
-------------
Upgrade to Couchbase SDK 2.6.0-beta.

For the default serializer, detect the presence of the
UnixMillisecondsConverter on an attribute to indicate that the attribute
is persisted as milliseconds.  For custom serializers, allow them to
implement IDateTimeSerializationFormatProvider to provide this
information.

When rendering UnixMilliseconds.ToDateTime (used internally for by
DateTimeComparisonExpressionTransformer), skip the STR_TO_MILLIS
conversion step for these attributes.

For DateTime parameters to N1QlFunctions methods, add a MILLIS_TO_STR
conversion for these attributes.

Refactor DocumentProxyDataMapper and BucketQueryExecutor so that they
can share a serializer.

Results
-------
Fields marked with UnixMillisecondsConverter are automatically supported
on queries for comparisons and date/time functions if using the default
serializer.  Custom serializers may implement
IDateTimeSerializationProvider to get the same support.