Issue #243: Allow ANSI NEST extents to be used in array subqueries
Motivation
----------
The resulting extent from an ANSI NEST is an array, which should be
usable in subqueries to apply additional filtering or projections.

Modifications
-------------
When parsing a subquery from clause, recognize a group join as a NEST
and treat it as an array extent instead of throwing an error.

Results
-------
Additional filters and projections can now be applied to the output of
ANSI NEST operations.