Issue #246: Implement more complete support for StringEnumConverter
Motivation
----------
When performing enum comparisons on enums serialized as a string using
StringEnumConverter, it is currently only supported if the JsonConverter
attribute is place on the Enum definition.  If placed on the property,
the comparison will not be rendered correctly in N1QL.

Modifications
-------------
Move the PartialEvaluatingExpressionTreeProcessor before the
SerializationExpressionTreeProcessor in the evaluation order.  This
causes expressions with a constant result around the enum to be
processed and simplified in advance of serialization expression
processing.  Changed from a NullEvaluatableExpressionFilter to
ExcludeSerializationConverterEvaluatableExpressionFilter to ensure that
serializaion conversions aren't simplified by the
PartialEvaluatingExpressionTreeProcessor.

Move the EnumComparisonExpressionTreeProcessor to execute prior to the
SerializationConverterExpressionTreeProcessor.  This ensures that the
comparison operation is between the enum types, not ints, so the
serialization converter requirement may be detected correctly.

Add nullable handling to EnumComparisonExpressionTreeProcessor.

Implement StringEnumSerializationConverter to detect and handle
rendering enum constants as strings rather than integers onto N1QL
queries.

Add optional injection of JsonConverter and MemberInfo to
SerializationConverter constructors, to provide necessary context to
StringEnumSerializationConverter.

Results
-------
Enum constants are rendered as strings when StringEnumConverter is
applied to the property or the Enum type and the constant is compared to
the property.

Other enum usages are unaffected.  Comparing one property with
StringEnumConverter to another property without StringEnumConverter is
still unsupported, only comparisons to constants.