Upgrade dependencies and TFMs for SDK 3
Motivation
----------
As part of SDK 3 compatibility, we need to update supported target
frameworks and some of our dependencies.

Modifications
-------------
Change the main library to target .NET Standard 2.0 and 2.1.

Change test libraries to target .NET Core 2.0 and 3.0.

Update version of dependencies to latest, as applicable, or to versions
that match SDK 3.

Remove targeting, dependencies, and files related to .NET Framework.
These are supplied via .NET Standard 2.0 in SDK 3.

Update to C# 8.

Results
-------
Better aligned for upgrade to SDK 3.

Closes #291