Add support for SingleAsync and SingleOrDefaultAsync
Motivation
----------
Support asynchronous queries which select only the first row and throw
an exception if there is more than one result.

Modifications
-------------
Create extension methods, an expression node, a result operator, and
streamed data classes to handle SingleAsync and SingleOrDefaultAsync.

Update N1QLQueryModelVisitor to recognize the new result operator.

Refactor Single and SingleOrDefault tests into separate files and add
tests for the async methods.

Results
-------
SingleAsync or SingleOrDefaultAsync may be called on IQueryable and they
function correctly so long as the IQueryable is from a
CollectionContext.

Relates to #281