Add support for CountAsync and LongCountASync
Motivation
----------
Support asynchronous queries which return counts.

Modifications
-------------
Create extension methods, an expression node, a result operator, and
streamed data classes to handle CountAsync and LongCountAsync.

Update N1QLQueryModelVisitor to recognize the new result operator.

Add tests for the async methods.

Results
-------
CountAsync or LongCountAsync may be called on IQueryable and they
function correctly so long as the IQueryable is from a
CollectionContext.

Relates to #281