Add support for AnyAsync and AllAsync
Motivation
----------
Support asynchronous queries which test a predicate against query
results.

Modifications
-------------
Create extension methods, expression nodes, and result operators for
AnyAsync and AllAsync.

Update N1QLQueryModelVisitor to recognize the new result operator.

Add tests for the async methods and reorganize tests for the sync
methods.

Results
-------
AnyAsync or AllAsync may be called on IQueryable and they function
correctly so long as the IQueryable is from a CollectionContext.

Relates to #281