Add support for SumAsync
Motivation
----------
Support asynchronous queries which return sums.

Modifications
-------------
Create extension methods, an expression node, and a result operator to
handle SumAsync.

Update N1QLQueryModelVisitor to recognize the new result operator.

Add tests for the async methods.

Results
-------
SumAsync may be called on IQueryable and will function correctly so long
as the IQueryable is from a CollectionContext.

Relates to #281