Add support for AverageAsync
Motivation
----------
Support asynchronous queries which return averages.

Modifications
-------------
Create extension methods, an expression node, and a result operator to
handle AverageAsync.

Update N1QLQueryModelVisitor to recognize the new result operator.

Add tests for the async methods.

Results
-------
AverageAsync may be called on IQueryable and will function correctly so
long as the IQueryable is from a CollectionContext.

Relates to #281