Switch back to BucketContext
Motivation
----------
After further consideration, BucketContext still makes sense for the
main context instead of CollectionContext. ICollectionQueryable still
works best for the actual IQueryable, since any given extent being
queried should be from a particular scope/collection.

By using BucketContext, if we reenable change tracking in the future
we'll be able to track changes across multiple scopes/collections and
submit them with a single method call. This was lost with the change
to CollectionContext.

Modifications
-------------
Switch CollectionContext back to BucketContext, and ICollectionContext
back to IBucketContext. For IBucketContext.Query, assume the default
collection.

Update tests and XML docs to match.

Results
-------
We're in a better position to add additional features in the future
which are backward compatible. This includes:

- Queries against collections other than the default collection, using
more specific method calls, or perhaps auto-filled properties on the
BucketContext (like DbSet in Entity Framework)
- Change tracking across the entire bucket

The API surface is also more backward compatible with 1.x of the LINQ
SDK, making upgrading easier.