Add QueryTimeout to IBucketContext
Motivation
----------
Allow the timeout for queries to be overridden from the cluster default
at the IBucketContext level.

Modifications
-------------
Add QueryTimeout to BucketContext and ClusterQueryExecutor.

Results
-------
When creating the BucketContext, possibly as part of the constructor
if inheriting, the timeout for queries can be overridden. If joining
between multiple BucketContexts, the timeout from the first referenced
BucketContext will apply.

Relates to #308