Add scan wait to ConsistentWith (#326)
Motivation
----------
When using ConsistentWith allow the consumer to specify a scan wait
limit.

Modifications
-------------
Add overload of ConsistentWith that accepts a scan wait TimeSpan, and
pass down through the expression node and clause. Apply this value
in ClusterQueryExecutor to the QueryOptions.

Results
-------
Scan wait can be controlled for queries performing RYOW using
ConsistentWith.