Add scan wait to ScanConsistency
Motivation
----------
When using ScanConsistency allow the consumer to specify a scan wait
limit.

Modifications
-------------
Add overload of ScanConsistency that accepts a scan wait TimeSpan, and
pass down through the expression node and clause. Apply this value
in ClusterQueryExecutor to the QueryOptions.

Results
-------
Scan wait can be controlled for queries performing with RequestPlus
consistency using ScanConsistency.

Close #308