Move DocumentFilterManager into cluster DI container
Motivation
----------
Make the DocumentFilterManager scoped to the cluster instead of a
static singleton. This will better support multiple clusters which may
have different requirements, as well as unit testing injection.

Modifications
-------------
Rewrite DocumentFilterManager to be an instance class instead of a
singleton. Perform some cleanup as well, such as making
DocumentFilterSet truly immutable, using nullable ref types, and fixing
up the XML doc comments.

Create extensions for ClusterOptions that register Linq2Couchbase on the
cluster services and allow extensibility.

Update some of the documentation with a modern example

Results
-------
DocumentFilterManager now lives on the ICluster.ClusterServices for
each cluster, with independent configuration.

To use Linq2Couchbase, AddLinq *must* be called on ClusterOptions prior
to connecting to the cluster.

Relates to #324