Drop extent naming
Motivation
----------
Since Couchbase Server 4.6.2 it isn't necessary for the extent name on
array subqueries to match the extent name used to make the index. The
minimum version is now 5.5, so this support is no longer necessary.

Modifications
-------------
Drop methods and classes related to controlling extent names in the
generated query.

Results
-------
A less confusing codebase and API surface.