Treat a string split as a query source for subqueries
Motivation
----------
Allow the application of filters like .Any() to the array that results
from a call to String.Split.

Modifications
-------------
Add an IQuerySourceExpressionNode for any call to String.Split so that
it may be treated as the source of an expression, and then handle this
as an array subquery in N1QLQueryModelVisitor.

Results
-------
`.Where(p => p.attr.Split().Any(q => q == 'x')` or similar queries will
now function.

Closes #337