Use unix milliseconds when sorting by date/times
Motivation
----------
Since our comparisons are based on unix milliseconds when comparing
DateTime properties, the same should be true when sorting by DateTime
properties. This is especially important for ensuring that indices
are useful, since the predicate and the order by may both need to
match the index key.

Modifications
-------------
Detect sorting by simple properties on documents which are ISO8601
strings and replace them with a STR_TO_MILLIS call.

Closes #331