Support using collections as query extents
Motivation
----------
Provide the back-end plumbing for querying against collections.

Modifications
-------------
Implement the query rendering logic that will format a query extent
which is a reference to a collection other than the default collection.

Add supporting unit tests.

Results
-------
This still doesn't add support for scopes/collections because they must
be exposed on the BucketContext. However, it is a key step in that
direction.

Relates to #354