Support for scope and collection identification
Motivation
----------
We need a way to annotate that a particular document resides in a
particular collection.

Modifications
-------------
Add the CouchbaseCollectionAttribute which annotates a POCO to indicate
the scope and collection of that POCO. If not present, assume the
document resides on the default collection.

Add tests and documentation.

Results
-------
POCOs controlled by the user may be targeted to a specific collection.
Some additional work is still required to allow extensibility for
overriding or POCOs not within the consumers control.

Relates to #354