Add IDocumentSet<T> support to BucketContext (#357)
Motivation
----------
Provide a cleaner mechanism for exposing strongly typed queryables on
classes inherited from BucketContext.

Allow the CouchbaseCollection attribute to be applied when it is either
not an option to add it to the POCO or the value on the POCO needs to be
overriden.

Modifications
-------------
Create the concept of a DocumentSet which mostly just exposes IQueryable
for a particular document type. It also accepts an overriden
scope/collection.

Add ContextMetadata with a cache to keep reflected information to build
a dynamic action that can quickly create DocumentSet instances and fill
the properties on the BucketContext. Use this cache whenever an
inherited BucketContext is created.

Make ICollectionQueryable interfaces internal, as they aren't used
externally and IDocumentSet is a cleaner way to expose the info to the
consumer.

Add a net5.0 target so we can use DynamicDependency to mark our
reflection from trimming. Note that this isn't the complete body of
work required to support member level trimming, but it's a start.

Results
-------
Consumers may now easily expose document sets as properties on their
BucketContext inherited classes. This will allow extensibility in the
future if support for other operations on IDocumentSet is desired.

Consumers may also use the IDocumentSet<T> properties to access the
backing ICouchbaseCollection to perform key/value operations. This
encourages a clean architecture where the CouchbaseCollection attribute
becomes the single point of truth for document storage location.

Consumers may now apply CouchbaseCollection attributes to documents in
a location other than on the POCO definition.